% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Create a new CSAS document based on a template}
\usage{
draft(
  type = c("resdoc", "fsar", "sr", "techreport"),
  directory = ".",
  edit = FALSE,
  create_rstudio_project = TRUE,
  ...
)
}
\arguments{
\item{type}{The type of document to draft. Must be one of \code{resdoc}, \code{fsar},
\code{sr}, or \code{techreport}.}

\item{directory}{The directory to place the draft document files.
Current directory by default}

\item{edit}{\code{TRUE} to edit the template immediately.}

\item{create_rstudio_project}{\code{TRUE} to create an RStudio project file.}

\item{...}{Other arguments to pass to \code{\link[rmarkdown:draft]{rmarkdown::draft()}}.}
}
\value{
A new project folder with template files for creating a csasdown
report.
}
\description{
Create a draft of an R Markdown CSAS document.
}
\details{
This is a light wrapper around \code{\link[rmarkdown:draft]{rmarkdown::draft()}}. Consult that
function for further details.
}
\examples{
\donttest{
# create a temporary example folder:
wd <- getwd()
example_path <- file.path(tempdir(), "csasdown-example")
dir.create(example_path)
setwd(example_path)

# create a draft template:
csasdown::draft("resdoc")

# return to original working directory after running example:
setwd(wd)

# clean up:
unlink(example_path, recursive = TRUE, force = TRUE)
}
}
\seealso{
\code{\link[=render]{render()}}
}
